+++ tsconfig.json
@@ -32,7 +32,6 @@
   "include": [
     "./**/*.ts",
     "./**/*.tsx",
-    "./next-env.d.ts",
     ".next/types/**/*.ts",
     ".next/dev/types/**/*.ts"
   ],
@@ -45,6 +44,7 @@
     "**/*.spec.ts",
     "**/*.spec.tsx",
     "./vitest.config.ts",
-    "./coverage"
+    "./coverage",
+    "./.next"
   ]
 }
